Try to separate console-specific code from general imgui code
