+++ crates/aether/src/ui/shaders/simple.hlsl
