+++ crates/aether/src/utils/mod.rs
@@ -4,6 +4,7 @@ pub mod debug_draw_3d;
 pub mod detour;
 pub mod fps_counter;
 pub mod geometry;
+pub mod imgui_ext;
 pub mod liang_barsky;
 pub mod marker;
 pub mod mem;
